fix(ontology): clarify telos gate receipts [impact-checked]
Moves typed-action payload harm classification under telos_gates ownership and marks default-only gate receipts as composite-derived instead of independent per-gate PASS values. Adds explicit logic-layer coverage for telos-required actions with post-default gate checks.

Verification: ruff check dharma_swarm/ontology.py dharma_swarm/telos_gates.py tests/test_ontology_telos_hardwire.py tests/test_logic_layer.py; pytest -q tests/test_ontology_telos_hardwire.py tests/test_ontology_registry.py tests/test_logic_layer.py::TestApplyAction tests/test_api.py::TestOntologyActions -q; make test-contracts; make test-hygiene; DHARMA_UPLIFT_ACK=impact-checked make uplift-guards
